Add the '-e' / '--no-err-files' option.
In some cases TESTDIR is read-only, so storing an .err file on a test
failure is impossible which make cram itself fail.

For example, if Cram is invoked as a part of 'make distcheck', which
deliberately runs a VPATH build with 'srcdir' and all its contents made
read-only, any test failure leads to Cram failing with a Python
traceback in the test log. Having a diff file instead would still be
handy to diagnose the failure.

With the new '-e' / '--no-err-files' option Cram does not try to create
.err files on test failures, but still prints the diff.

+++ cram/_cli.py
@@ -54,7 +54,8 @@ def _patch(cmd, diff):
     out, retcode = execute([cmd, '-p0'], stdin=b('').join(diff))
     return retcode == 0
 
-def runcli(tests, quiet=False, verbose=False, patchcmd=None, answer=None):
+def runcli(tests, quiet=False, verbose=False, patchcmd=None, answer=None,
+           noerrfiles=False):
     """Run tests with command line interface input/output.
 
     tests should be a sequence of 2-tuples containing the following:
@@ -102,12 +103,13 @@ def runcli(tests, quiet=False, verbose=False, patchcmd=None, answer=None):
                 if not quiet:
                     _log('\n', None, verbose)
 
-                errfile = open(errpath, 'wb')
-                try:
-                    for line in postout:
-                        errfile.write(line)
-                finally:
-                    errfile.close()
+                if not noerrfiles:
+                    errfile = open(errpath, 'wb')
+                    try:
+                        for line in postout:
+                            errfile.write(line)
+                    finally:
+                        errfile.close()
 
                 if not quiet:
                     origdiff = diff
@@ -120,7 +122,8 @@ def runcli(tests, quiet=False, verbose=False, patchcmd=None, answer=None):
                         _prompt('Accept this change?', 'yN', answer) == 'y'):
                         if _patch(patchcmd, diff):
                             _log(None, path + b(': merged output\n'), verbose)
-                            os.remove(errpath)
+                            if not noerrfiles:
+                                os.remove(errpath)
                         else:
                             _log(path + b(': merge failed\n'))
